更改数据库使用由ADO.NET为Entity Framework，同时修复一些Bug

+++ NeoMapleStory/Game/Client/MapleSkinColor.cs
@@ -26,7 +26,7 @@
 
         public byte ColorId { get; private set; }
 
-        public static MapleSkinColor GetByColorId(int colorId)
+        public static MapleSkinColor GetByColorId(byte colorId)
         {
             switch (colorId)
             {
