+++ NeoMapleStory/Game/Skill/SkillEntry.cs
@@ -2,14 +2,14 @@
 {
     public class SkillEntry
     {
-        public SkillEntry(int skillevel, int masterlevel)
+        public SkillEntry(byte skillevel, byte masterlevel)
         {
             SkilLevel = skillevel;
             MasterLevel = masterlevel;
         }
 
-        public int SkilLevel { get; }
-        public int MasterLevel { get; }
+        public byte SkilLevel { get; }
+        public byte MasterLevel { get; }
 
         public override string ToString() => $"{SkilLevel}:{MasterLevel}";
     }
